requirements.txt
requests==2.24.0
Flask==1.1.2
Flask_MySQLdb==0.2.0
pandas==1.0.5
PyYAML==5.3.1
